Moved backup block class and improved imports
