+++ minecraft_model_reader/api/__init__.py
@@ -1,0 +1,4 @@
+try:
+    from amulet.api.block import Block
+except ModuleNotFoundError:
+    from minecraft_model_reader.api.amulet.block import Block
